create integration test for product service client and server
